pyproject.toml
[tool.poetry]
name = "leylinechallenge"
version = "0.1.0"
description = ""
authors = ["ecopes <emacopes@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
utils = "^1.0.2"
djangorestframework = "^3.15.2"
pillow = "^10.3.0"
redis = "^5.0.7"
django-cors-headers = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
